Rename checkin controller
Credential name will be less confusing once the volunteer utility is
written
Factor out QR code image utilities
